+++ CHANGELOG.md
@@ -9,6 +9,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 [Unreleased]: https://github.com/atomist/rug/compare/0.15.0...HEAD
 
+### Fixed
+
+-   Testing editors no longer crashes when provided invalid
+    parameters [#444][444]
+
+[444]: https://github.com/atomist/rug/issues/444
+
 ## [0.15.0] - 2017-03-17
 
 [0.15.0]: https://github.com/atomist/rug/compare/0.14.0...0.15.0